Migrated new `search.js` to typescript
